requirements.txt
python-dotenv==0.21.0
requests==2.28.1
schedule==1.1.0
selenium==4.5.0
webdriver-manager==3.8.3
